+++ app/Providers/RouteServiceProvider.php
@@ -70,7 +70,7 @@ class RouteServiceProvider extends ServiceProvider
      */
     protected function mapAdminRoutes()
     {
-        Route::middleware(['web', 'auth'])
+        Route::middleware(['web', 'admin'])
              ->namespace($this->namespace . '\Admin')
              ->prefix('admin')
              ->group(base_path('routes/admin.php'));