+++ CHANGELOG.md
@@ -35,6 +35,7 @@ changes made when rewriting.
 ### Fixed
 - Clearing or truncating chat now properly checks for permission level
 - Clearing or truncating error log now properly checks for permission level
+- Fixed Laravel admin controllers not logging the user visit
 - Fixed missing private message icon in admin overview
 
 ## [0.6.0] - 2018-08-26